future/requirements.txt
asgiref==3.3.4
bracex==2.1.1
casbin==1.1.1
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
simpleeval==0.9.10
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
wcmatch==8.2
